+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](http://semver.org/).
   of the setting
 - Dup notify opts before mutating (#345)
 ### Changed
+- Breadcrumbs on by default
 - Added Faktory plugin -@scottrobertson
 
 ## [4.5.6] - 2020-01-08